Rename start-api.js to api-dom.js for relevancy
